Add desync research related symbols
Symbols that start with "?_dr" and are followed by three digits are related to the desync research. 
There are some "bad" values that cause dolphin to fall off-sync when playing against wii online, and most of these "?_dr" functions take those values as input for some operation and propagate the error. 
Source of the error is still unknown.